+++ src/Sentry/SentryOptions.cs
@@ -1370,6 +1370,8 @@ public class SentryOptions
         );
 
         NetworkStatusListener = new PollingNetworkStatusListener(this);
+
+        Experimental = new ExperimentalSentryOptions(this);
     }
 
     /// <summary>
@@ -1905,4 +1907,55 @@ public class SentryOptions
         "ServiceStack",
         "Java.Interop",
     ];
+
+    /// <summary>
+    /// Sentry features that are currently in an experimental state.
+    /// </summary>
+    /// <remarks>
+    /// Experimental features are subject to binary, source and behavioral breaking changes in future updates.
+    /// </remarks>
+    public ExperimentalSentryOptions Experimental { get; }
+
+    /// <summary>
+    /// Sentry features that are currently in an experimental state.
+    /// </summary>
+    /// <remarks>
+    /// Experimental features are subject to binary, source and behavioral breaking changes in future updates.
+    /// </remarks>
+    public class ExperimentalSentryOptions
+    {
+        private readonly SentryOptions _options;
+        private SentryTraceMetricsCallbacks? _beforeSendMetric;
+
+        internal ExperimentalSentryOptions(SentryOptions options)
+        {
+            _options = options;
+        }
+
+        internal SentryTraceMetricsCallbacks? BeforeSendMetricInternal => _beforeSendMetric;
+
+        /// <summary>
+        /// When set to <see langword="false"/>, the SDK does not generate and send metrics to Sentry via <see cref="SentrySdk"/>.
+        /// Defaults to <see langword="true"/>.
+        /// </summary>
+        /// <seealso href="https://develop.sentry.dev/sdk/telemetry/metrics/"/>
+        public bool EnableMetrics { get; set; } = true;
+
+        /// <summary>
+        /// Sets a callback function to be invoked before sending the metric of numeric type <typeparamref name="T"/> to Sentry.
+        /// When the delegate throws an <see cref="Exception"/> during invocation, the metric will not be captured.
+        /// </summary>
+        /// <typeparam name="T">The numeric type of the metric.</typeparam>
+        /// <remarks>
+        /// It can be used to modify the metric object before being sent to Sentry.
+        /// To prevent the metric from being sent to Sentry, return <see langword="null"/>.
+        /// Supported numeric value types for <typeparamref name="T"/> are <see langword="byte"/>, <see langword="short"/>, <see langword="int"/>, <see langword="long"/>, <see langword="float"/>, and <see langword="double"/>.
+        /// </remarks>
+        /// <seealso href="https://develop.sentry.dev/sdk/telemetry/metrics/"/>
+        public void SetBeforeSendMetric<T>(Func<SentryMetric<T>, SentryMetric<T>?> beforeSendMetric) where T : struct
+        {
+            _beforeSendMetric ??= new SentryTraceMetricsCallbacks();
+            _beforeSendMetric.Set(beforeSendMetric, _options.DiagnosticLogger);
+        }
+    }
 }